Inner:修复EnumerateTags在非递归场景下只遍历schemaNS问题

+++ frameworks/innerkitsimpl/test/unittest/image_source_test/image_source_test.cpp
@@ -230,7 +230,7 @@ public:
         return bufferPtr_;
     }
 
-    uint32_t GetStreamSize()
+    size_t GetStreamSize()
     {
         return size_;
     }
